Start to pull apart title menus into a base-class/implementation system
