+++ src/frontend.h
@@ -34,10 +34,10 @@ enum tMode
 	TUTORIAL,		// 5  tutorial/fastplay
 	CREDITS,		// 6  credits
 	PROTOCOL,		// 7  MULTIPLAYER, select proto
-	MULTIOPTION,	// 8 MULTIPLAYER, select game options
+	OBS_MULTI_OPT,	// 8 MULTIPLAYER, select game options - DO NOT USE, now a WzTitleUI thing, see src/titleui/titleui.h -- 20kdc
 	FORCESELECT,	// 9 MULTIPLAYER, Force design screen
 	GAMEFIND,		// 10 MULTIPLAYER, gamefinder.
-	MULTILIMIT,		// 11 MULTIPLAYER, Limit the multistuff.
+	OBS_MULLIM,		// 11 MULTIPLAYER, Limit the multistuff - DO NOT USE, now a WzTitleUI thing, see src/titleui/titleui.h -- 20kdc
 	STARTGAME,		// 12 Fire up the game
 	SHOWINTRO,		// 13 reshow the intro
 	QUIT,			// 14 leaving game
@@ -50,7 +50,6 @@ enum tMode
 	CAMPAIGNS,              // 21 campaign selector
 };
 
-extern tMode titleMode;					// the global case
 extern tMode lastTitleMode;
 
 #define MAX_LEVEL_NAME_SIZE	(256)
@@ -73,7 +72,7 @@ bool runVideoOptionsMenu();
 bool runMouseOptionsMenu();
 bool runTutorialMenu();
 
-void addTopForm();
+void addTopForm(bool wide);
 void addBottomForm();
 void addBackdrop();
 void addTextButton(UDWORD id, UDWORD PosX, UDWORD PosY, const std::string &txt, unsigned int style);
