+++ src/multiplay.cpp
@@ -75,6 +75,7 @@
 #include "multiint.h"
 #include "keymap.h"
 #include "cheat.h"
+#include "main.h"								// for gamemode
 
 // ////////////////////////////////////////////////////////////////////////////
 // ////////////////////////////////////////////////////////////////////////////
@@ -1502,7 +1503,7 @@ bool recvTextMessage(NETQUEUE queue)
 	eventFireCallbackTrigger((TRIGGER_TYPE)CALL_AI_MSG);
 
 	// make some noise!
-	if (titleMode == MULTIOPTION || titleMode == MULTILIMIT)
+	if (GetGameMode() != GS_NORMAL)
 	{
 		audio_PlayTrack(FE_AUDIO_MESSAGEEND);
 	}
